[tag] chore: move get_media_path_from_file to media module
* Update references in cli for `media-tagger` and `media-similarity`
