+++ libs/media_similarity/media_similarity/__init__.py
@@ -30,4 +30,4 @@ __all__ = [
   'MediaClusteringRequest',
 ]
 
-__version__ = '0.7.1'
+__version__ = '0.7.2'
