+++ libs/media_tagging/media_tagging/__init__.py
@@ -28,4 +28,4 @@ __all__ = [
   'MediaTaggingService',
   'MediaTaggingRequest',
 ]
-__version__ = '2.2.0'
+__version__ = '2.2.1'
